backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
gunicorn>=22.0
psycopg2-binary>=2.9
django-filter>=24.3
djangorestframework-simplejwt>=5.3
